fix tests and linters
